feat: unified filters + default all time in date
