+++ .mcp.json
@@ -1,0 +1,9 @@
+{
+  "mcpServers": {
+    "puppeteer": {
+      "command": "npx",
+      "args": ["-y", "puppeteer-mcp-server"],
+      "env": {}
+    }
+  }
+}
